[filter] Supports time string with a unit suffix in the maxIdle options of muxers

+++ src/filters/mux.hpp
@@ -114,7 +114,7 @@ private:
 
   class SessionManager : public pjs::RefCount<SessionManager> {
     struct Options {
-      int max_idle = 10;
+      double max_idle = 10;
     };
 
     SessionManager(MuxBase *mux)
